separate settings
Create a folder for settings constants and setting tab component
